1. ResponseEntity in REST controller 2. Catch EmptyResultSetException, throw custom DeviceNotFoundException 3. RabbitMQ communication. Message receiver.

+++ src/main/java/com/vanchutin/model/Device.java
@@ -8,13 +8,13 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public @Data class Device {
 
-    public Device(int id, String name, Status status){
+    public Device(long id, String name, Status status){
         this.id = id;
         this.name = name;
         this.status = status;
     }
 
-    private Integer id;
+    private Long id;
 
     private String name;
